UnitTests tested on both Windows and Linux, TStream class modified to handle some problems with different data types on Windows and Linux

+++ Source/UnitTests/PersistenceExample/ClassA.h
@@ -15,7 +15,7 @@ class ClassA : public BaseClass
 			
 		TShortString DeviceName;
 		TShortString DeviceType;
-		unsigned long Flags;
+		unsigned int Flags;
